refactor: entity간 양방향 참조 제거하는 방향으로 진행중

+++ src/test/java/com/kustaurant/kustaurant/comment/domain/PostCommentTest.java
@@ -1,8 +1,8 @@
 package com.kustaurant.kustaurant.comment.domain;
 
-import com.kustaurant.kustaurant.comment.domain.PostComment;
-import com.kustaurant.kustaurant.post.enums.ContentStatus;
-import com.kustaurant.kustaurant.post.enums.ReactionStatus;
+import com.kustaurant.kustaurant.post.comment.domain.PostComment;
+import com.kustaurant.kustaurant.post.post.enums.ContentStatus;
+import com.kustaurant.kustaurant.post.post.enums.ReactionStatus;
 import org.junit.jupiter.api.Test;
 
 import java.time.LocalDateTime;
@@ -15,7 +15,7 @@ class PostCommentTest {
     void 댓글_생성시_기본값이_잘_설정된다() {
         // Given
         String content = "Test comment";
-        Integer userId = 1;
+        Long userId = 1L;
         Integer postId = 10;
 
         // When
@@ -36,7 +36,7 @@ class PostCommentTest {
     @Test
     void 댓글_좋아요_처음_누르면_좋아요가_증가한다() {
         // Given
-        PostComment comment = PostComment.create("Like test", 1, 1);
+        PostComment comment = PostComment.create("Like test", 1L, 1);
 
         // When
         ReactionStatus result = comment.toggleLike(false, false);
@@ -51,7 +51,7 @@ class PostCommentTest {
     @Test
     void 댓글_좋아요_취소하면_좋아요가_감소한다() {
         // Given
-        PostComment comment = PostComment.create("Like cancel", 1, 1);
+        PostComment comment = PostComment.create("Like cancel", 1L, 1);
         comment.increaseLikeCount(1);
 
         // When
@@ -66,7 +66,7 @@ class PostCommentTest {
     @Test
     void 댓글_싫어요에서_좋아요로_전환하면_싫어요가_감소하고_좋아요가_증가한다() {
         // Given
-        PostComment comment = PostComment.create("Switch", 1, 1);
+        PostComment comment = PostComment.create("Switch", 1L, 1);
         comment.increaseDislikeCount(1);
 
         // When
@@ -82,8 +82,8 @@ class PostCommentTest {
     @Test
     void 댓글을_삭제하면_상태가_DELETED로_변경된다() {
         // Given
-        PostComment parentComment = PostComment.create("Parent comment", 1, 1);
-        PostComment reply = PostComment.create("Reply", 1, 1);
+        PostComment parentComment = PostComment.create("Parent comment", 1L, 1);
+        PostComment reply = PostComment.create("Reply", 1L, 1);
         parentComment.getReplies().add(reply);
 
         // When
@@ -97,7 +97,7 @@ class PostCommentTest {
     @Test
     void 시간차이를_분단위로_계산할_수_있다() {
         // Given
-        PostComment comment = PostComment.create("Comment", 1, 1);
+        PostComment comment = PostComment.create("Comment", 1L, 1);
         comment.setCreatedAt(LocalDateTime.now().minusMinutes(10));
 
         // When
@@ -110,7 +110,7 @@ class PostCommentTest {
     @Test
     void 시간차이를_초단위로_계산할_수_있다() {
         // Given
-        PostComment comment = PostComment.create("Comment", 1, 1);
+        PostComment comment = PostComment.create("Comment", 1L, 1);
         comment.setCreatedAt(LocalDateTime.now().minusSeconds(30));
 
         // When
